fix(webpage-builder): restore add block modal trigger
